# [SYNCOPE-1689] Consolidate Provision, Mapping and Items into single JSON column