--- a/client/idm/console/src/main/java/org/apache/syncope/client/console/commons/IdMStatusProvider.java
+++ b/client/idm/console/src/main/java/org/apache/syncope/client/console/commons/IdMStatusProvider.java
@@ -30,7 +30,7 @@ import org.apache.syncope.client.console.status.ReconStatusUtils;
 import org.apache.syncope.client.console.wicket.markup.html.form.ActionLink;
 import org.apache.syncope.client.ui.commons.status.ConnObjectWrapper;
 import org.apache.syncope.common.lib.to.AnyTO;
-import org.apache.syncope.common.lib.to.ConnObjectTO;
+import org.apache.syncope.common.lib.to.ConnObject;
 import org.apache.syncope.common.lib.types.IdMEntitlement;
 
 public class IdMStatusProvider implements StatusProvider {
@@ -38,7 +38,7 @@ public class IdMStatusProvider implements StatusProvider {
     private static final long serialVersionUID = 1875374599950896631L;
 
     @Override
-    public Optional<Pair<ConnObjectTO, ConnObjectTO>> get(
+    public Optional<Pair<ConnObject, ConnObject>> get(
             final String anyTypeKey, final String connObjectKeyValue, final String resource) {
 
         return ReconStatusUtils.getReconStatus(anyTypeKey, connObjectKeyValue, resource).
@@ -46,12 +46,12 @@ public class IdMStatusProvider implements StatusProvider {
     }
 
     @Override
-    public List<Triple<ConnObjectTO, ConnObjectWrapper, String>> get(
+    public List<Triple<ConnObject, ConnObjectWrapper, String>> get(
             final AnyTO any, final Collection<String> resources) {
 
         return ReconStatusUtils.getReconStatuses(
                 any.getType(), any.getKey(), any.getResources()).stream().
-                map(status -> Triple.<ConnObjectTO, ConnObjectWrapper, String>of(
+                map(status -> Triple.<ConnObject, ConnObjectWrapper, String>of(
                 status.getRight().getOnSyncope(),
                 new ConnObjectWrapper(any, status.getLeft(), status.getRight().getOnResource()),
                 null)).