--- a/core/persistence-api/src/main/java/org/apache/syncope/core/persistence/api/dao/DAO.java
+++ b/core/persistence-api/src/main/java/org/apache/syncope/core/persistence/api/dao/DAO.java
@@ -25,6 +25,4 @@ public interface DAO<E extends Entity> {
     void refresh(E entity);
 
     void detach(E entity);
-
-    void clear();
 }