… client/idm/console/src/main/java/org/apache/syncope/client/console/wizards/resources/ConnectorDetailsPanel.java
… client/idm/console/src/main/java/org/apache/syncope/client/console/wizards/resources/ConnectorWizardBuilder.java
… client/idm/console/src/main/java/org/apache/syncope/client/console/wizards/resources/ResourceProvision.java
… client/idrepo/common-ui/src/main/java/org/apache/syncope/client/ui/commons/status/ConnObjectWrapper.java
… client/idrepo/common-ui/src/main/java/org/apache/syncope/client/ui/commons/status/StatusUtils.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/commons/IdRepoStatusProvider.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/commons/StatusProvider.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/panels/AnyDirectoryPanel.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/panels/Realm.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wicket/extensions/markup/html/repeater/data/table/ConnObjectAttrColumn.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/widgets/ItemTransformerWidget.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/widgets/JEXLTransformerWidget.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wizards/any/ConnObjectPanel.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wizards/any/StatusPanel.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wizards/mapping/AbstractMappingPanel.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wizards/mapping/ItemTransformersTogglePanel.java
… client/idrepo/console/src/main/java/org/apache/syncope/client/console/wizards/mapping/JEXLTransformersTogglePanel.java
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyItemTransformer.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyLogicActions.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyPropagationActions.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyProvisionSorter.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyPullActions.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyPullCorrelationRule.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyPushCorrelationRule.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MyReportlet.groovy
… client/idrepo/console/src/main/resources/org/apache/syncope/client/console/implementations/MySchedTaskJobDelegate.groovy
… common/am/lib/src/main/java/org/apache/syncope/common/lib/to/AttrRepoTO.java
… common/am/lib/src/main/java/org/apache/syncope/common/lib/to/AuthModuleTO.java
… common/idm/lib/src/main/java/org/apache/syncope/common/lib/to/ConnIdBundle.java
… common/idm/lib/src/main/java/org/apache/syncope/common/lib/to/ConnIdObjectClass.java