The purpose of this PR is to explore the possibility to consolidate a set of JPA entities from their own tables into single JSON column.

The idea is that such strategy should improve the general provisioning performance, especially during propagation, as no joins will be needed to fetch the required information.

At least some comparison tests shall be performed against current `master` branch and this PR to understand if the proposed change is effective or not.

It could be an idea to do the same with Connector and Resource capabilities.