@ilgrosso 2022-07-14
Very interesting results from some smoke performance tests.

I've been using [ilgrosso/syncopeRestClient](https://github.com/ilgrosso/syncopeRestClient) to create users with propagation against Syncope instances started via `mvn -Pdebug,all` under `fit/core-reference` directory.

| | `master` | this PR |
|--------------|-----------|------------|
| 1000 users | 1145426 ms (~19 minutes)  | 189939 ms (~3 minutes) |